Change view components to separate html and ts
