+++ webapp-giftlist-1.0/src/views/BookedGifts/BookedGifts.vue
@@ -1,0 +1,2 @@
+<script src="./booked-gifts.ts"></script>
+<template src="./booked-gifts.html"></template>
