+++ webapp-giftlist-1.0/src/views/BookedGifts/booked-gifts.html
@@ -1,0 +1,10 @@
+<DefaultLayout :title="labels.titles.booked">
+    <div class="grid grid-cols-4 gap-4">
+        <BookedGift v-for="gift in bookedGifts" :key="gift.id" :gift="gift" @click="handleDetailsModal(gift)" />
+    </div>
+
+    <Modal :show="modal.showModal" :title="modal.title" :confirmText="modal.confirmText" :cancelText="modal.cancelText"
+        @close="modal.showModal = false" @confirm="modal.confirm">
+        <GiftDetails v-if="selectedGift" :gift="selectedGift" />
+    </Modal>
+</DefaultLayout>
