+++ webapp-giftlist-1.0/src/views/Friends/Friends.vue
@@ -1,0 +1,8 @@
+<script src="./friends.ts"></script>
+<template src="./friends.html"></template>
+<style lang="scss" scoped>
+.size {
+	max-height: 75vh;
+	overflow: hidden;
+}
+</style>
