+++ webapp-giftlist-1.0/src/views/Friends/friends.ts
@@ -1,0 +1,73 @@
+import { defineComponent, ref } from "vue";
+
+import labels from "@/labels/fr/labels.json";
+
+import DefaultLayout from "@/components/Styled/DefaultLayout.vue";
+import Button from "@/components/Styled/Button.vue";
+import Subtitle from "@/components/Styled/Subtitle.vue";
+import FormInputText from "@/components/Inputs/FormInputText.vue";
+
+import { CheckCircleIcon, XCircleIcon } from "@heroicons/vue/outline";
+
+export default defineComponent({
+    name: "Friends",
+    components: { Button, DefaultLayout, FormInputText, Subtitle, CheckCircleIcon, XCircleIcon },
+    setup() {
+        const acceptFriendRequest = () => {
+            console.log("Friends.vue - acceptFriendRequest");
+        };
+
+        const declineFriendRequest = () => {
+            console.log("Friends.vue - declineFriendRequest");
+        };
+
+        const deleteFriend = () => {
+            console.log("Friends.vue - deleteFriend");
+        };
+
+        const emailInput = ref({
+            label: labels.friends.addFriendsInput.label,
+            placeholder: labels.friends.addFriendsInput.placeholder,
+            value: "",
+            errorMessage: "",
+            helperText: labels.friends.addFriendsInput.helperText,
+        });
+
+        const handleEmailInputChange = (value: string) => {
+            emailInput.value.value = value;
+        };
+
+        const sendFriendRequest = () => {
+            console.log("Friends.vue - sendFriendRequest");
+        };
+
+        const emailInvitationInput = ref({
+            label: labels.friends.inviteInput.label,
+            placeholder: labels.friends.inviteInput.placeholder,
+            value: "",
+            errorMessage: "",
+            helperText: labels.friends.inviteInput.helperText,
+        });
+
+        const handleEmailInvitationInputChange = (value: string) => {
+            emailInvitationInput.value.value = value;
+        };
+
+        const sendInvitationRequest = () => {
+            console.log("Friends.vue - sendInvitationRequest");
+        };
+
+        return {
+            labels,
+            acceptFriendRequest,
+            declineFriendRequest,
+            deleteFriend,
+            emailInput,
+            handleEmailInputChange,
+            sendFriendRequest,
+            emailInvitationInput,
+            handleEmailInvitationInputChange,
+            sendInvitationRequest,
+        };
+    },
+});
