+++ webapp-giftlist-1.0/src/views/Gift/Gift.vue
@@ -1,0 +1,2 @@
+<script src="./gift.ts"></script>
+<template src="./gift.html"></template>
