+++ webapp-giftlist-1.0/src/views/Home/Home.vue
@@ -1,0 +1,2 @@
+<script src="./home.ts"></script>
+<template src="./home.html"></template>
