+++ webapp-giftlist-1.0/src/views/Home/home.html
@@ -1,0 +1,24 @@
+<DefaultLayout title="Home">
+    <div>
+        <Subtitle>Resources</Subtitle>
+        <ul>
+            <li>
+                <a href="https://heroicons.com/" rel="noreferrer noopener" class="text-blue-500">Heroicons</a>
+            </li>
+            <li>
+                <a href="https://tailwindui.com/" rel="noreferrer noopener" class="text-blue-500">TailwindUI</a>
+            </li>
+            <li>
+                <a href="https://headlessui.dev/vue" rel="noreferrer noopener" class="text-blue-500">HeadlessUI</a>
+            </li>
+            <li>
+                <a href="https://auth0.com/docs/api/management/v2#!/Users/get_users" rel="noreferrer noopener"
+                    class="text-blue-500">Auth0 Management API Docs</a>
+            </li>
+            <li>
+                <a href="https://auth0.com/docs/security/tokens/access-tokens/get-management-api-access-tokens-for-production"
+                    rel="noreferrer noopener" class="text-blue-500">Auth0 Management API Get Token</a>
+            </li>
+        </ul>
+    </div>
+</DefaultLayout>
