+++ webapp-giftlist-1.0/src/views/Home/home.ts
@@ -1,0 +1,9 @@
+import { defineComponent } from "vue";
+
+import DefaultLayout from "@/components/Styled/DefaultLayout.vue";
+import Subtitle from "@/components/Styled/Subtitle.vue";
+
+export default defineComponent({
+    name: "Home",
+    components: { DefaultLayout, Subtitle },
+});
