+++ webapp-giftlist-1.0/src/views/List/List.vue
@@ -1,0 +1,2 @@
+<script src="./list.ts"></script>
+<template src="./list.html"></template>
