+++ webapp-giftlist-1.0/src/views/ListSettings/ListSettings.vue
@@ -1,0 +1,2 @@
+<script src="./list-settings.ts"></script>
+<template src="./list-settings.html"></template>
