+++ webapp-giftlist-1.0/src/views/ListSettings/list-settings.html
@@ -1,0 +1,61 @@
+<DefaultLayout :title="labels.titles.listSettings" back>
+    <div class="w-10/12 mx-auto">
+        <Disclosure v-slot="{ open }">
+            <DisclosureButton @click="handleOpen('general')" class="
+                    flex
+                    justify-between
+                    w-full
+                    px-4
+                    py-4
+                    text-sm
+                    font-medium
+                    text-left text-indigo-900
+                    bg-indigo-100
+                    rounded-lg
+                    hover:bg-indigo-200
+                    focus:outline-none
+                    focus-visible:ring focus-visible:ring-indigo-500 focus-visible:ring-opacity-75
+                ">
+                <span>{{ labels.listOptions.generalTitle }}</span>
+                <ChevronUpIcon :class="open ? 'transform rotate-180' : ''" class="w-5 h-5 text-indigo-500" />
+            </DisclosureButton>
+            <transition enter-active-class="transition duration-300 ease-out" enter-from-class="transform opacity-0"
+                enter-to-class="transform opacity-100">
+                <DisclosurePanel v-show="openGeneral" static class="px-4 pt-4 pb-2 text-sm text-black">
+                    <ListFormStep1 :values="generalInformation" @change="handleGeneralInformationChange" />
+                </DisclosurePanel>
+            </transition>
+        </Disclosure>
+        <Disclosure as="div" class="mt-2" v-slot="{ open }">
+            <DisclosureButton @click="handleOpen('share')" class="
+                    flex
+                    justify-between
+                    w-full
+                    px-4
+                    py-4
+                    text-sm
+                    font-medium
+                    text-left text-indigo-900
+                    bg-indigo-100
+                    rounded-lg
+                    hover:bg-indigo-200
+                    focus:outline-none
+                    focus-visible:ring focus-visible:ring-indigo-500 focus-visible:ring-opacity-75
+                ">
+                <span>{{ labels.listOptions.sharingTitle }}</span>
+                <ChevronUpIcon :class="open ? 'transform rotate-180' : ''" class="w-5 h-5 text-indigo-500" />
+            </DisclosureButton>
+            <transition enter-active-class="transition duration-300 ease-out" enter-from-class="transform opacity-0"
+                enter-to-class="transform opacity-100">
+                <DisclosurePanel v-show="openShare" static class="px-4 pt-4 pb-2 text-sm text-black">
+                    <ListFormStep2 :values="sharingOptions" @change="handleSharingOptionsChange" />
+                </DisclosurePanel>
+            </transition>
+        </Disclosure>
+
+        <div class="flex justify-end gap-4 mt-4">
+            <Button btnStyle="danger" @click="cancel">Retour</Button>
+            <Button btnStyle="primary" @click="saveChanges">Enregistrer les changements</Button>
+        </div>
+    </div>
+</DefaultLayout>
