+++ webapp-giftlist-1.0/src/views/Lists/Lists.vue
@@ -1,0 +1,2 @@
+<script src="./Lists.ts"></script>
+<template src="./Lists.html"></template>
