+++ webapp-giftlist-1.0/src/views/NewGift/NewGift.vue
@@ -1,0 +1,2 @@
+<script src="./new-gift.ts"></script>
+<template src="./new-gift.html"></template>
