+++ webapp-giftlist-1.0/src/views/NewGift/new-gift.html
@@ -1,0 +1,4 @@
+<DefaultLayout :title="labels.titles.newGift" back>
+    <GiftForm action="create" :values="giftInformation" @change="handleGiftInformationChange" @cancel="cancel"
+        @confirm="createGift" />
+</DefaultLayout>
