+++ webapp-giftlist-1.0/src/views/NewList/NewList.vue
@@ -1,0 +1,2 @@
+<script src="./new-list.ts"></script>
+<template src="./new-list.html"></template>
