+++ webapp-giftlist-1.0/src/views/NewList/new-list.html
@@ -1,0 +1,44 @@
+<DefaultLayout :title="labels.titles.newList" back>
+    <div class="flex flex-col">
+        <div class="relative my-4 flex flex-col border rounded-lg">
+            <Subtitle class="absolute top-0 left-4 transform -translate-y-1/2 bg-white px-4">
+                {{ stepTitle }}
+            </Subtitle>
+
+            <component class="p-4 my-4 rounded-md" :is="currentComponent" :values="listInformation['step' + step]"
+                @change="handleListInformationChange">
+            </component>
+        </div>
+
+        <div class="flex justify-between">
+            <div class="flex gap-4">
+                <Button btnStyle="danger" hasIcon @click="cancel">
+                    <template v-slot:icon>
+                        <XIcon />
+                    </template>
+                    {{ labels.newList.buttons.cancel }}
+                </Button>
+                <Button btnStyle="secondary" hasIcon v-show="step > 1" @click="step--">
+                    <template v-slot:icon>
+                        <ArrowLeftIcon />
+                    </template>
+                    {{ labels.newList.buttons.previous }}
+                </Button>
+            </div>
+            <div class="flex gap-4">
+                <Button :btnStyle="step === maxStep ? 'primary' : 'secondary'" hasIcon @click="skipToList">
+                    <template v-slot:icon>
+                        <CheckIcon />
+                    </template>
+                    {{ labels.newList.buttons.confirm }}
+                </Button>
+                <Button btnStyle="primary" hasIcon v-show="step != maxStep" @click="nextAction">
+                    <template v-slot:icon>
+                        <ArrowRightIcon />
+                    </template>
+                    {{ nextButtonText }}
+                </Button>
+            </div>
+        </div>
+    </div>
+</DefaultLayout>
