--- webapp-giftlist-1.0/src/views/Profile.vue
@@ -1,175 +1,0 @@
-<template>
-	<DefaultLayout :title="labels.titles.profile" back>
-		<div class="flex flex-col my-4">
-			<div class="flex items-center justify-between border rounded-lg px-4 my-4">
-				<img :src="auth.user.picture" alt="Profile Image" class="rounded-full m-4" />
-
-				<div class="mx-4 w-1/4">
-					<Subtitle>{{ auth.user.nickname }}</Subtitle>
-					<div>{{ labels.profile.email }} {{ auth.user.email }}</div>
-					<button
-						v-if="!auth.user.email_verified"
-						class="text-indigo-600"
-						@click="verifyEmail"
-					>
-						{{ labels.profile.verifyEmail }}
-					</button>
-				</div>
-				<div class="w-1/4">
-					<div class="flex">
-						<span
-							v-for="(friend, i) in friends"
-							:key="friend.id"
-							class="
-								w-12
-								h-12
-								bg-gray-300
-								rounded-full
-								grid
-								place-items-center
-								font-bold
-								text-sm text-black
-								border border-white
-								shadow-sm
-							"
-							:class="i > 0 ? '-ml-2' : null"
-							>{{ friend.name }}</span
-						>
-						<span
-							class="
-								w-12
-								h-12
-								bg-gray-300
-								rounded-full
-								grid
-								place-items-center
-								font-bold
-								text-sm text-black
-								border border-white
-								shadow-sm
-								-ml-2
-							"
-							>8+</span
-						>
-					</div>
-					<div class="pl-1">12 {{ labels.profile.friends }}</div>
-					<router-link to="/app/profile/friends" class="ml-1 mt-4 text-indigo-600">
-						{{ labels.profile.manageFriends }}
-					</router-link>
-				</div>
-				<div class="w-1/4">
-					<div class="flex items-center justify-between mb-4">
-						<div class="flex flex-col items-center">
-							<span>
-								<strong> 3 </strong>
-							</span>
-							<span class="text-center leading-tight">
-								{{ labels.profile.createdLists }}
-							</span>
-						</div>
-						<div class="flex flex-col items-center">
-							<span>
-								<strong> 12 </strong>
-							</span>
-							<span class="text-center leading-tight">
-								{{ labels.profile.createdGifts }}
-							</span>
-						</div>
-						<div class="flex flex-col items-center">
-							<span>
-								<strong> 5 </strong>
-							</span>
-							<span class="text-center leading-tight">
-								{{ labels.profile.bookedGifts }}
-							</span>
-						</div>
-					</div>
-				</div>
-			</div>
-
-			<div class="flex items-stretch divide-x">
-				<div class="flex-1 p-8">
-					<div class="pb-5">
-						<Subtitle>{{ labels.profile.editEmailTitle }}</Subtitle>
-						<div class="mt-2 mb-4">
-							<p>
-								{{ labels.profile.editEmailCurrent }} <strong>{{ auth.user.email }}</strong>
-							</p>
-							<p>
-								{{ labels.profile.editEmailText }}
-							</p>
-						</div>
-						<Button btnStyle="secondary" class="w-1/2" @click="changeEmail">
-							{{ labels.profile.editEmailButton }}
-						</Button>
-					</div>
-					<div class="pt-5">
-						<Subtitle>{{ labels.profile.editPasswordTitle }}</Subtitle>
-						<p class="mt-2 mb-4">
-							{{ labels.profile.editPasswordText }}
-						</p>
-						<Button btnStyle="secondary" class="w-1/2" @click="changePassword">
-							{{ labels.profile.editPasswordButton }}
-						</Button>
-					</div>
-				</div>
-				<div class="flex-1 flex flex-col p-8">
-					<Subtitle>{{ labels.profile.dataTitle }}</Subtitle>
-					<p class="mt-2 mb-4">
-						{{ labels.profile.dataText }}
-					</p>
-					<Button btnStyle="secondary" class="w-1/2" @click="downloadData">
-						{{ labels.profile.dataButton }}
-					</Button>
-				</div>
-			</div>
-		</div>
-	</DefaultLayout>
-</template>
-
-<script lang="ts">
-import { defineComponent, inject, ref } from "vue";
-
-import labels from "@/labels/fr/labels.json";
-
-import Button from "@/components/Styled/Button.vue";
-import DefaultLayout from "@/components/Styled/DefaultLayout.vue";
-import Subtitle from "@/components/Styled/Subtitle.vue";
-
-export default defineComponent({
-	name: "Profile",
-	components: { Button, DefaultLayout, Subtitle },
-	setup() {
-		const auth = ref(inject("Auth") as any);
-		const friends = [
-			{ id: 0, name: "ND" },
-			{ id: 1, name: "ML" },
-			{ id: 2, name: "PC" },
-			{ id: 3, name: "ML" },
-		];
-
-		const verifyEmail = () => {
-			console.log("Profile.vue - verifyEmail");
-		};
-		const changeEmail = () => {
-			console.log("Profile.vue - changeEmail");
-		};
-		const changePassword = () => {
-			console.log("Profile.vue - changePassword");
-		};
-		const downloadData = () => {
-			console.log("Profile.vue - downloadData");
-		};
-
-		return {
-			labels,
-			auth,
-			friends,
-			verifyEmail,
-			changeEmail,
-			changePassword,
-			downloadData,
-		};
-	},
-});
-</script>
