+++ webapp-giftlist-1.0/src/views/Profile/Profile.vue
@@ -1,0 +1,2 @@
+<script src="./profile.ts"></script>
+<template src="./profile.html"></template>
