+++ webapp-giftlist-1.0/src/views/Profile/profile.ts
@@ -1,0 +1,44 @@
+import { defineComponent, inject, ref } from "vue";
+
+import labels from "@/labels/fr/labels.json";
+
+import Button from "@/components/Styled/Button.vue";
+import DefaultLayout from "@/components/Styled/DefaultLayout.vue";
+import Subtitle from "@/components/Styled/Subtitle.vue";
+
+export default defineComponent({
+    name: "Profile",
+    components: { Button, DefaultLayout, Subtitle },
+    setup() {
+        const auth = ref(inject("Auth") as any);
+        const friends = [
+            { id: 0, name: "ND" },
+            { id: 1, name: "ML" },
+            { id: 2, name: "PC" },
+            { id: 3, name: "ML" },
+        ];
+
+        const verifyEmail = () => {
+            console.log("Profile.vue - verifyEmail");
+        };
+        const changeEmail = () => {
+            console.log("Profile.vue - changeEmail");
+        };
+        const changePassword = () => {
+            console.log("Profile.vue - changePassword");
+        };
+        const downloadData = () => {
+            console.log("Profile.vue - downloadData");
+        };
+
+        return {
+            labels,
+            auth,
+            friends,
+            verifyEmail,
+            changeEmail,
+            changePassword,
+            downloadData,
+        };
+    },
+});
