+++ webapp-giftlist-1.0/src/views/Settings/Settings.vue
@@ -1,0 +1,2 @@
+<script src="./settings.ts"></script>
+<template src="./settings.html"></template>
