+++ webapp-giftlist-1.0/src/views/SharedList/SharedList.vue
@@ -1,0 +1,2 @@
+<script src="./shared-list.ts"></script>
+<template src="./shared-list.html"></template>
