+++ webapp-giftlist-1.0/src/views/SharedLists/SharedLists.vue
@@ -1,0 +1,2 @@
+<script src="./shared-lists.ts"></script>
+<template src="./shared-lists.html"></template>
